+++ lista07/ex02/manifest.mf
