+++ lista07/ex02/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=/home/gustavoscarenci/snap/netbeans/69/build.properties
